Fix some broken tests stories around orders table and form
