--- src/components/Office/OrdersDetailForm.module.scss
@@ -1,15 +1,0 @@
-@import '../../shared/styles/basics';
-@import '../../shared/styles/mixins';
-@import '../../shared/styles/colors';
-@import '../../shared/styles/custom';
-
-.OrdersDetailForm {
-  label {
-    @include u-margin-bottom(1);
-    color: $base-darker;
-  }
-
-  :global .DayPickerInput {
-    width: 100%;
-  }
-}
