+++ src/components/Office/OrdersDetailForm/OrdersDetailForm.module.scss
@@ -1,0 +1,15 @@
+@import 'shared/styles/_basics';
+@import 'shared/styles/_mixins';
+@import 'shared/styles/colors';
+@import 'shared/styles/_custom';
+
+.OrdersDetailForm {
+  label {
+    @include u-margin-bottom(1);
+    color: $base-darker;
+  }
+
+  :global .DayPickerInput {
+    display: block;
+  }
+}
